+++ scripts/add_missing_data_to_vcf.py
@@ -30,45 +30,48 @@ with open(vcf, "r") as infile:
 
 outstring=vcf+"_missing"+str(missing)+".vcf"
 
-## header line is #CHROM
-## there are going to be 10 header data before the rest of the data
+if os.isfile(outstring) == True:
+	print("SKIPPING")
+else:
+	## header line is #CHROM
+	## there are going to be 10 header data before the rest of the data
 
-possiblebases=["A","C","T","G","0","1"]
-replacelines=copy.deepcopy(lines)
-
-numinds=None
-nummiss=None
-ntodraw=None
-for i in list(range(len(lines))):
-	#print(i)
-	thisline = lines[i]
-	firstchar = thisline[0]
-	if firstchar != "#":
-		## do stuff
-		split = thisline.split("\t")
-		data = split[9:]
-		newdata=copy.deepcopy(data)
-		if(numinds==None):
-			numinds=len(data)
-		if(nummiss==None):
-			nummiss=int(numinds*missing)
-		if(ntodraw==None):
-			ntodraw=list(range(0,numinds,1))
-		thismiss=random.sample(ntodraw,nummiss)
-		thismiss.sort()
+	possiblebases=["A","C","T","G","0","1"]
+	replacelines=copy.deepcopy(lines)
+	
+	numinds=None
+	nummiss=None
+	ntodraw=None
+	for i in list(range(len(lines))):
+		#print(i)
+		thisline = lines[i]
+		firstchar = thisline[0]
+		if firstchar != "#":
+			## do stuff
+			split = thisline.split("\t")
+			data = split[9:]
+			newdata=copy.deepcopy(data)
+			if(numinds==None):
+				numinds=len(data)
+			if(nummiss==None):
+				nummiss=int(numinds*missing)
+			if(ntodraw==None):
+				ntodraw=list(range(0,numinds,1))
+			thismiss=random.sample(ntodraw,nummiss)
+			thismiss.sort()
 		
-		for index in thismiss:
-			dat2sub = newdata[index]
-			for x in possiblebases:
-				dat2sub = dat2sub.replace(x,".")
-			newdata[index] = dat2sub
-		newline = split[:9]+newdata
-		newline = "\t".join(newline)
-	else:
-		newline = thisline
-	replacelines[i] = newline
-
-with open(outstring,"w") as outfile:
-	outfile.writelines(replacelines)
+			for index in thismiss:
+				dat2sub = newdata[index]
+				for x in possiblebases:
+					dat2sub = dat2sub.replace(x,".")
+				newdata[index] = dat2sub
+			newline = split[:9]+newdata
+			newline = "\t".join(newline)
+		else:
+			newline = thisline
+		replacelines[i] = newline
+		
+	with open(outstring,"w") as outfile:
+		outfile.writelines(replacelines)
